fastmcp_server/requirements.txt
fastmcp>=2.0,<3
requests>=2.31,<3
pydantic>=2.6,<3
python-dotenv>=1.0,<2
